refactor(ci): split content-hash into types, inputs, markers, hash
content-hash.ts mixed path tables, success markers, cache keys, and git
digest collection. Extract identity vocabulary, JOB_CONTENT_INPUTS, and
markers; keep hashing/cache keys/git digests in content-hash.ts. Public
exports from scripts/ci-routing.ts unchanged.
