+++ scripts/ci-routing/content-hash.ts
@@ -3,7 +3,7 @@
  * early-exit success after restoring a validated success marker.
  *
  * Path routing (./routing) decides whether a job is scheduled. This module
- * owns hash identity, success markers, and cache keys.
+ * owns hashing, cache keys, and git digest collection.
  *
  * - Hashes are fail-closed: every matched input path must have a digest.
  * - Expanding JOB_CONTENT_INPUTS changes the patterns string inside the digest
@@ -14,325 +14,20 @@
  *   job is scheduled.
  * - Invalidation: bump CONTENT_HASH_SCHEMA, edit a matched input, change
  *   JOB_CONTENT_INPUTS patterns, or set repo variable CI_DISABLE_CONTENT_HASH=1.
+ *
+ * Path tables: content-hash-inputs.ts. Markers: content-hash-markers.ts.
+ * Identity: content-hash-types.ts.
  */
 import { createHash } from "node:crypto";
 
-import { matchPathPattern, type ChangeFlags, type PlanOptions } from "./routing";
-
-// Schema 2: split former monolithic `build` into build / svelte_check / docs_site.
-export const CONTENT_HASH_SCHEMA = 2;
-
-/**
- * Physical CI executions that may content-hash short-circuit.
- * Distinct from JobName: component is four shards (svelte chromium, svelte
- * firefox+webkit, spikes, journeys); consumer is matrixed.
- */
-export type CacheableExecution =
-  | "packages_dist"
-  | "unit"
-  | "build"
-  | "svelte_check"
-  | "docs_site"
-  | "actions_security"
-  | "bench_smoke"
-  | "interaction_perf"
-  | "component_svelte"
-  | "component_svelte_fx"
-  | "component_spikes"
-  | "component_journeys"
-  | "consumer";
-
-export const CACHEABLE_EXECUTIONS: readonly CacheableExecution[] = [
-  "packages_dist",
-  "unit",
-  "build",
-  "svelte_check",
-  "docs_site",
-  "actions_security",
-  "bench_smoke",
-  "interaction_perf",
-  "component_svelte",
-  "component_svelte_fx",
-  "component_spikes",
-  "component_journeys",
-  "consumer",
-] as const;
-
-/**
- * Toolchain + recipe files folded into every hashable execution so a workflow
- * or router change cannot reuse results produced under a different recipe.
- */
-const UNIVERSAL_CONTENT_INPUTS: readonly string[] = [
-  ".github/workflows/ci.yml",
-  // Domain job bodies live in reusable workflows (issue #392).
-  ".github/workflows/ci-*.yml",
-  // Composite actions hold the success-marker protocol after extraction from ci.yml.
-  ".github/actions/**",
-  "scripts/ci-routing.ts",
-  "scripts/ci-routing.test.ts",
-  "scripts/ci-routing/**",
-  "bun.lock",
-  "package.json",
-  "tsconfig.json",
-  "tsconfig.base.json",
-  "bunfig.toml",
-];
-
-/**
- * Docs app + every script that can change svelte-check / vite docs-site
- * outcomes (generators, $scripts imports, pages-link / metadata gates).
- * Shared so the three post-split jobs cannot false-green on omitted inputs.
- */
-const DOCS_SURFACE_CONTENT_INPUTS: readonly string[] = [
-  "apps/docs/**",
-  "examples/**",
-  "lifecycle.json",
-  // apps/docs package.json build/check invoke these generators.
-  "scripts/gen-docs-routes.ts",
-  "scripts/docs-route-inventory.ts",
-  "scripts/docs-route-inventory.test.ts",
-  "scripts/gen-docs-search.ts",
-  "scripts/gen-docs-search.test.ts",
-  "scripts/gen-gallery-previews.ts",
-  "scripts/gen-gallery-previews.test.ts",
-  "scripts/gallery-preview-provenance.ts",
-  "scripts/gallery-preview-provenance.test.ts",
-  "scripts/check-docs-metadata.ts",
-  "scripts/check-pages-links.ts",
-  // #784: build/check --check gen-lesson-charts; build runs docs-csp after vite.
-  "scripts/gen-lesson-charts.ts",
-  "scripts/gen-theme-static-shells.ts",
-  "scripts/docs-csp.ts",
-  // $scripts imports used by docs routes / layout (typecheck + published site).
-  "scripts/gen-llms.ts",
-  "scripts/gen-llms.test.ts",
-  "scripts/llms-markdown.ts",
-  "scripts/llms-guide-content.ts",
-  "scripts/docs-seo.ts",
-  "scripts/diagnostic-docs.ts",
-  "scripts/cli-docs.ts",
-  "scripts/guide-code-contract.ts",
-  "scripts/quickstart.ts",
-  "scripts/quickstart/**",
-  "scripts/highlight-code.ts",
-  "scripts/highlight-code.test.ts",
-  "scripts/deployment-artifact.ts",
-  "scripts/deployment-smoke.ts",
-  "scripts/deployment-smoke-cli.ts",
-];
-
-/**
- * Conservative content inputs per physical execution.
- * Patterns use the same matcher as path routing (no negation). Prefer broad
- * trees over incomplete maps that could false-green.
- */
-export const JOB_CONTENT_INPUTS: Record<CacheableExecution, readonly string[]> = {
-  packages_dist: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-  ],
-  unit: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "benchmarks/**",
-    "scripts/**",
-    "tests/evals/**",
-    // workers/** (when present) run in the unit suite.
-    "workers/**",
-    "docs/accessibility/**",
-    ".github/ISSUE_TEMPLATE/**",
-    ".github/DISCUSSION_TEMPLATE/**",
-    ".changeset/**",
-    "skills/**",
-    "lifecycle.json",
-    "apps/docs/**",
-    "examples/**",
-    ".github/workflows/**",
-    // scripts/actionlint.test.ts asserts self-hosted labels against this file.
-    ".github/actionlint.yaml",
-    "support-matrix.json",
-    ".oxlintrc.json",
-    ".oxfmtrc.json",
-    ".markdownlint-cli2.jsonc",
-    "knip.jsonc",
-    ".pre-commit-config.yaml",
-    // scripts/readme-showcase.test.ts contracts the published README snippets.
-    "README.md",
-  ],
-  // Package build + knip + type-aware + publint + examples tsc (no vite docs site).
-  // apps/docs stays hashed: knip + oxlint --type-aware still cover the docs app,
-  // and routing still schedules build for docsSurface changes (Codex P2).
-  build: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "apps/docs/**",
-    "examples/**",
-    // check:scripts-ci-routing tsc covers scripts/ci-routing/** (#734); the
-    // rest of scripts/** is still type-aware lint + knip only.
-    "scripts/**",
-    "tests/evals/**",
-    // build re-enters `bun run check` plus repo-wide type-aware lint + knip.
-    "workers/**",
-    "skills/**",
-    "lifecycle.json",
-    "support-matrix.json",
-    ".oxlintrc.json",
-    ".oxfmtrc.json",
-    ".markdownlint-cli2.jsonc",
-    "knip.jsonc",
-    ".pre-commit-config.yaml",
-    ".github/workflows/**",
-  ],
-  // svelte-check for packages/svelte + apps/docs (generators + check).
-  svelte_check: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    ...DOCS_SURFACE_CONTENT_INPUTS,
-  ],
-  // Full vite adapter-static docs site + packed pages-link gate.
-  docs_site: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    ...DOCS_SURFACE_CONTENT_INPUTS,
-  ],
-  actions_security: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    ".github/workflows/**",
-    ".github/actionlint.yaml",
-    "scripts/actionlint.ts",
-    "scripts/actionlint.test.ts",
-  ],
-  bench_smoke: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "benchmarks/**",
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-  ],
-  interaction_perf: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "tests/performance/**",
-    "apps/docs/src/routes/__perf/**",
-    "benchmarks/interaction-budgets.json",
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "apps/docs/**",
-    "examples/**",
-  ],
-  component_svelte: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "skills/ggsvelte/**",
-  ],
-  // Same inputs as component_svelte (shared packages/svelte tree); distinct
-  // execution key so chromium+coverage and firefox+webkit cache independently.
-  component_svelte_fx: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "skills/ggsvelte/**",
-  ],
-  component_spikes: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "spikes/**",
-  ],
-  component_journeys: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "apps/docs/**",
-    "examples/**",
-    "tests/visual/**",
-    "skills/ggsvelte/**",
-    "lifecycle.json",
-    // Docs generators that schedule docs_journeys must bust this cache
-    // (Codex P2: content-only scripts were omitting gen-docs-search / gallery).
-    "scripts/gen-llms.ts",
-    "scripts/gen-llms.test.ts",
-    "scripts/llms-markdown.ts",
-    // llms-markdown imports highlight-code for fenced-block HTML; hash it or
-    // a highlight-only change can cache-hit and skip docs journeys (Codex P2).
-    "scripts/highlight-code.ts",
-    "scripts/highlight-code.test.ts",
-    "scripts/llms-guide-content.ts",
-    "scripts/docs-seo.ts",
-    "scripts/diagnostic-docs.ts",
-    "scripts/cli-docs.ts",
-    "scripts/guide-code-contract.ts",
-    "scripts/gen-docs-search.ts",
-    "scripts/gen-docs-search.test.ts",
-    "scripts/gen-gallery-previews.ts",
-    "scripts/gen-gallery-previews.test.ts",
-    "scripts/gallery-preview-provenance.ts",
-    "scripts/gallery-preview-provenance.test.ts",
-    // Journeys assert lesson-chart img counts on /guide/getting-started (#784).
-    "scripts/gen-lesson-charts.ts",
-    "scripts/gen-theme-static-shells.ts",
-    "scripts/cloudflare-pages-config.test.ts",
-    "scripts/deployment-artifact.ts",
-    "scripts/deployment-artifact.test.ts",
-    "scripts/deployment-smoke.ts",
-    "scripts/deployment-smoke-cli.ts",
-    "scripts/deployment-smoke.test.ts",
-    "scripts/quickstart.ts",
-    "scripts/quickstart/**",
-    "scripts/gen-manifest.ts",
-    "scripts/gen-lifecycle.ts",
-  ],
-  consumer: [
-    ...UNIVERSAL_CONTENT_INPUTS,
-    "packages/spec/**",
-    "packages/core/**",
-    "packages/svelte/**",
-    "scripts/consumer-compat.ts",
-    "scripts/consumer-compat.test.ts",
-    "scripts/guide-code-contract.ts",
-    "scripts/quickstart.ts",
-    "scripts/quickstart/**",
-    "scripts/quickstart-timing.ts",
-    "scripts/quickstart-timing.test.ts",
-    "scripts/support-matrix.ts",
-    "scripts/support-matrix.test.ts",
-    "support-matrix.json",
-  ],
-};
-
-export function listJobContentPaths(
-  execution: CacheableExecution,
-  allPaths: readonly string[],
-): string[] {
-  const patterns = JOB_CONTENT_INPUTS[execution];
-  const matched = new Set<string>();
-  for (const raw of allPaths) {
-    const file = raw.replaceAll("\\", "/").replace(/^\.\//, "");
-    if (!file || file === ".") continue;
-    for (const pattern of patterns) {
-      if (matchPathPattern(pattern, file)) {
-        matched.add(file);
-        break;
-      }
-    }
-  }
-  return [...matched].toSorted();
-}
+import { type ChangeFlags, type PlanOptions } from "./routing";
+import {
+  CONTENT_HASH_SCHEMA,
+  formatShardKeyPart,
+  type CacheableExecution,
+  type ExecutionShard,
+} from "./content-hash-types";
+import { JOB_CONTENT_INPUTS, listJobContentPaths } from "./content-hash-inputs";
 
 export type HashJobInputsOptions = {
   /** Override schema (tests only). Production always uses CONTENT_HASH_SCHEMA. */
@@ -399,19 +94,6 @@ export function shouldBypassContentCache(changes: ChangeFlags, options: PlanOpti
   if (options.forceAll === true) return true;
   return changes.lockfile || changes.ci_workflow || changes.ci_routing || changes.ci_actions;
 }
-/**
- * One leg of a matrixed fan-out (vitest / playwright `--shard=index/total`).
- * 1-based `index`, matching both runners' CLI convention.
- *
- * Shard identity is part of cache identity: the legs hash the same content but
- * execute disjoint test files, so a marker written by one leg must never
- * satisfy another. `total` is included so re-shaping the fan-out (3 → 4 legs)
- * misses every old marker instead of reusing legs that now cover other files.
- */
-export type ExecutionShard = {
-  index: number;
-  total: number;
-};
 
 export type ContentHashCacheKeyInput = {
   execution: CacheableExecution;
@@ -466,90 +148,6 @@ export function contentHashCacheKey(input: ContentHashCacheKeyInput): string {
 
 function sanitizeKeyPart(value: string): string {
   return value.replaceAll(/[^a-zA-Z0-9._-]+/g, "");
-}
-
-/** `shard1of3` — stable in both cache keys and marker filenames. */
-function formatShardKeyPart(shard: ExecutionShard): string {
-  return `shard${shard.index}of${shard.total}`;
-}
-
-function parseShard(value: unknown): ExecutionShard | null {
-  if (value === null || typeof value !== "object") return null;
-  const obj = value as Record<string, unknown>;
-  const index = obj["index"];
-  const total = obj["total"];
-  if (!Number.isInteger(index) || !Number.isInteger(total)) return null;
-  const i = index as number;
-  const t = total as number;
-  if (i < 1 || t < 1 || i > t) return null;
-  return { index: i, total: t };
-}
-
-function sameShard(a: ExecutionShard | undefined, b: ExecutionShard | undefined): boolean {
-  if (a === undefined || b === undefined) return a === undefined && b === undefined;
-  return a.index === b.index && a.total === b.total;
-}
-
-export type SuccessMarker = {
-  schema: number;
-  execution: CacheableExecution;
-  hash: string;
-  /** Present only for legs of a sharded execution. */
-  shard?: ExecutionShard;
-};
-
-export function serializeSuccessMarker(marker: SuccessMarker): string {
-  return `${JSON.stringify(marker)}\n`;
-}
-
-export function parseSuccessMarker(body: string): SuccessMarker | null {
-  try {
-    const parsed: unknown = JSON.parse(body);
-    if (parsed === null || typeof parsed !== "object") return null;
-    const obj = parsed as Record<string, unknown>;
-    if (typeof obj["schema"] !== "number") return null;
-    if (typeof obj["execution"] !== "string") return null;
-    if (typeof obj["hash"] !== "string" || obj["hash"].length === 0) return null;
-    if (!(CACHEABLE_EXECUTIONS as readonly string[]).includes(obj["execution"])) return null;
-    const base = {
-      schema: obj["schema"],
-      execution: obj["execution"] as CacheableExecution,
-      hash: obj["hash"],
-    };
-    if (obj["shard"] === undefined) return base;
-    // A shard field that is present but unreadable is a corrupt marker, not an
-    // unsharded one — fail closed rather than validating against the wrong leg.
-    const shard = parseShard(obj["shard"]);
-    if (shard === null) return null;
-    return { ...base, shard };
-  } catch {
-    return null;
-  }
-}
-
-export function validateSuccessMarker(
-  marker: SuccessMarker | null,
-  expected: {
-    execution: CacheableExecution;
-    hash: string;
-    schema?: number;
-    shard?: ExecutionShard;
-  },
-): boolean {
-  if (marker === null) return false;
-  const schema = expected.schema ?? CONTENT_HASH_SCHEMA;
-  return (
-    marker.schema === schema &&
-    marker.execution === expected.execution &&
-    marker.hash === expected.hash &&
-    sameShard(marker.shard, expected.shard)
-  );
-}
-
-/** Success-marker path relative to repo root (actions/cache path). */
-export function successMarkerPath(execution: CacheableExecution, shard?: ExecutionShard): string {
-  const leg = shard === undefined ? "" : `-${shard.index}of${shard.total}`;
-  return `.ci-content-hash/${execution}${leg}.ok`;
 }
 
 /** Parse one `git ls-tree -r` line into mode/type/oid/path. */
